1st modification to sizeof

+++ 0x00-hello_world/6-size.c
@@ -14,10 +14,10 @@ int main(void)
 	char c;
 	float d;
 
-	printf("The size of an int is: %lu.\n", sizeof(a));
-	printf("The size of a double is: %lu.\n", sizeof(b));
-	printf("the size of a char is: %lu.\n", sizeof(c));
-	printf("The size of a float is: %lu.\", sizeof(d));
+	printf("The size of an int is: %lu.\n", (unsigned long)sizeof(a));
+	printf("The size of a double is: %lu.\n", (unsigned long)sizeof(b));
+	printf("the size of a char is: %lu.\n", (unsigned long)sizeof(c));
+	printf("The size of a float is: %lu.\", (unsigned long)sizeof(d));
 
 	return (0);
 }
